Remove OSSF scorecard workflow
Scorecard is now part of the test workflow. (it's in the same spirit as
testing/linting/auditing)

We don't bother running on the branch-protection event, though because
how often is that even happening. And the recurring cron schedule would
run often enough to update the relevant scorecard metric. (and on top of
the frequency of the runs, the corresponding metrics are so nearly
insignificant that (in)frequency of re-checking is no big deal)